% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsePhyloProfile.R
\name{createRootedTreeCr}
\alias{createRootedTreeCr}
\title{Create rooted tree from a taxonomy matrix}
\usage{
createRootedTreeCr(df, rootTaxon = NULL)
}
\arguments{
\item{df}{data frame contains taxonomy matrix used for generating tree
(see distDf in example)}

\item{rootTaxon}{taxon used for rooting the taxonomy tree}
}
\value{
A rooted taxonomy tree as an object of class "phylo".
}
\description{
Create rooted tree from a taxonomy matrix
}
\author{
Vinh Tran {tran@bio.uni-frankfurt.de}
}
