% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsePhyloProfile.R
\name{getInputTaxaNameCr}
\alias{getInputTaxaNameCr}
\title{Get NCBI taxon names for a selected list of taxa}
\usage{
getInputTaxaNameCr(rankName, taxonIDs = NULL)
}
\arguments{
\item{rankName}{taxonomy rank (e.g. "species","phylum",...)}

\item{taxonIDs}{list of taxon IDs (e.g. ncbi1234). Default = NULL}
}
\value{
Data frame contains a list of full names, taxonomy ranks and parent
IDs for the input taxa.
}
\description{
Get NCBI taxon names from
"PhyloProfile/data/taxonNamesReduced.txt" for a list of input taxa
}
\author{
Vinh Tran {tran@bio.uni-frankfurt.de}
}
