% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsePhyloProfile.R
\name{getNameListCr}
\alias{getNameListCr}
\title{Get list of pre-installed NCBI taxon names}
\usage{
getNameListCr()
}
\value{
List of taxon IDs, their full names, taxonomy ranks and parent IDs
obtained from "PhyloProfile/data/taxonNamesReduced.txt"
}
\description{
Get all NCBI taxon names from
"PhyloProfile/data/taxonNamesReduced.txt"
}
\author{
Vinh Tran {tran@bio.uni-frankfurt.de}
}
