% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsePhyloProfile.R
\name{getTaxonomyMatrixCr}
\alias{getTaxonomyMatrixCr}
\title{Get taxonomy matrix}
\usage{
getTaxonomyMatrixCr(subsetTaxaCheck = FALSE, taxonIDs = NULL)
}
\arguments{
\item{subsetTaxaCheck}{TRUE/FALSE subset taxonomy matrix based on input taxon
IDs. Default = FALSE.}

\item{taxonIDs}{list of input taxon IDs (e.g. ncbi1234). Default = NULL.}
}
\value{
Data frame contains the (subset of) taxonomy matrix for list of
input taxa.
}
\description{
Get the (full or subset) taxonomy matrix from
"data/taxonomyMatrix.txt" based on an input taxon list
}
\author{
Vinh Tran {tran@bio.uni-frankfurt.de}
}
