% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runPhyloprofile.R
\name{runPhyloCellulase}
\alias{runPhyloCellulase}
\title{Run PhyloCellulase app}
\usage{
runPhyloCellulase()
}
\value{
A shiny application - GUI version of PhyloProfile
}
\description{
Run PhyloCellulase app
}
