% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsePhyloProfile.R
\name{sortInputTaxaCr}
\alias{sortInputTaxaCr}
\title{Sort list of (super)taxa based on a selected reference (super)taxon}
\usage{
sortInputTaxaCr(taxonIDs = NULL, rankName, refTaxon = NULL,
    taxaTree = NULL)
}
\arguments{
\item{taxonIDs}{list of taxon IDs (e.g.: ncbi1234, ncbi9999, ...). Default =
NULL.}

\item{rankName}{working taxonomy rank (e.g. "species", "phylum",...)}

\item{refTaxon}{selected reference taxon. Default = NULL.}

\item{taxaTree}{taxonomy tree for the input taxa (optional). Default = NULL.}
}
\value{
A taxonomy matrix for the input taxa ordered by the selected
reference taxon. This matrix is sorted either based on the NCBI taxonomy
info, or based on an user-defined taxonomy tree (if provided).
}
\description{
Sort list of (super)taxa based on a selected reference (super)taxon
}
\author{
Vinh Tran {tran@bio.uni-frankfurt.de}
}
